backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
supabase==2.18.0
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
python-dotenv==1.1.1
pydantic>=2.11.7,<3.0.0
httpx==0.28.1
pydantic-settings==2.10.1
